setup/glue/requirements.txt
pyspark==3.5.0
py4j==0.10.9.7
boto3==1.34.0
snowflake-connector-python==3.6.0
pyarrow==14.0.1
pandas==2.1.4
pyyaml==6.0.1